requirements.txt
requests
sqlalchemy
sqlalchemy-postgres-copy
pandas
psycopg2-binary
fire
flask
flask-sqlalchemy
geoalchemy